+++ src/common/Microsoft.DotNet.UpgradeAssistant.Abstractions/IUpgradeReadyCheck.cs
@@ -17,11 +17,17 @@ namespace Microsoft.DotNet.UpgradeAssistant
         string Id { get; }
 
         /// <summary>
+        /// Gets a message that tells users what they can do if their project is not ready to upgrade.
+        /// </summary>
+        string UpgradeMessage { get; }
+
+        /// <summary>
         /// Verifies that given the supplied project, an upgrade is possible.
         /// </summary>
         /// <param name="project">The project to be upgraded.</param>
+        /// <param name="options">Options for readiness checks.</param>
         /// <param name="token">A cancellation token.</param>
         /// <returns><c>true</c> if the project is ready, <c>false</c> if it is not.</returns>
-        Task<bool> IsReadyAsync(IProject project, CancellationToken token);
+        Task<UpgradeReadiness> IsReadyAsync(IProject project, UpgradeReadinessOptions options, CancellationToken token);
     }
 }