+++ src/components/Microsoft.DotNet.UpgradeAssistant/Checks/WcfServerCheck.cs
@@ -1,0 +1,74 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Globalization;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
+
+namespace Microsoft.DotNet.UpgradeAssistant.Checks
+{
+    /// <summary>
+    /// Users running upgrade-assistant on projects that build WCF Services
+    /// should be informed that this is an unsupported technology before they
+    /// invest significant time running the tool.
+    /// </summary>
+    public class WcfServerCheck : IUpgradeReadyCheck
+    {
+        private const string UPGRADE_LINK = "https://aka.ms/migrate-wcf-to-grpc";
+        private const string CATEGORY = "WCF Server-side Services";
+        private const string WCF_MESSAGE = "Support for {0} is limited to .NET Full Framework. To learn more please read: {1}";
+        private readonly ILogger<WcfServerCheck> _logger;
+
+        public WcfServerCheck(ILogger<WcfServerCheck> logger)
+        {
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+        }
+
+        /// <summary>
+        /// Gets the value 'WcfServerCheck'.
+        /// </summary>
+        public string Id => nameof(WcfServerCheck);
+
+        public string UpgradeMessage => string.Format(CultureInfo.InvariantCulture, WCF_MESSAGE, CATEGORY, UPGRADE_LINK);
+
+        public Task<UpgradeReadiness> IsReadyAsync(IProject project, UpgradeReadinessOptions options, CancellationToken token)
+        {
+            if (project is null)
+            {
+                throw new ArgumentNullException(nameof(project));
+            }
+
+            if (options is null)
+            {
+                throw new ArgumentNullException(nameof(options));
+            }
+
+            if (options.IgnoreUnsupportedFeatures)
+            {
+                return Task.FromResult(UpgradeReadiness.Ready);
+            }
+
+            // are there any svc.cs files in this project?
+            if (project.FindFiles($".svc{GetExtensionForLanguage(project)}").Any())
+            {
+                _logger.LogError(WCF_MESSAGE, CATEGORY, UPGRADE_LINK);
+                return Task.FromResult(UpgradeReadiness.Unsupported);
+            }
+
+            return Task.FromResult(UpgradeReadiness.Ready);
+        }
+
+        private static string GetExtensionForLanguage(IProject project)
+        {
+            if (project.Language == Language.VisualBasic)
+            {
+                return ".vb";
+            }
+
+            return ".cs";
+        }
+    }
+}