+++ src/steps/Microsoft.DotNet.UpgradeAssistant.Steps.Solution/CurrentProjectSelectionStep.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 
 namespace Microsoft.DotNet.UpgradeAssistant.Steps.Solution
 {
@@ -17,6 +18,7 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Solution
         private readonly IUserInput _input;
         private readonly ITargetFrameworkMonikerComparer _tfmComparer;
         private readonly ITargetFrameworkSelector _tfmSelector;
+        private readonly IOptions<UpgradeReadinessOptions> _upgradeOptions;
         private IProject[]? _orderedProjects;
 
         public override IEnumerable<string> DependsOn { get; } = new[]
@@ -35,6 +37,7 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Solution
             IUserInput input,
             ITargetFrameworkMonikerComparer tfmComparer,
             ITargetFrameworkSelector tfmSelector,
+            IOptions<UpgradeReadinessOptions> upgradeOptions,
             ILogger<CurrentProjectSelectionStep> logger)
             : base(logger)
         {
@@ -42,6 +45,12 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Solution
             _input = input ?? throw new ArgumentNullException(nameof(input));
             _tfmComparer = tfmComparer ?? throw new ArgumentNullException(nameof(tfmComparer));
             _tfmSelector = tfmSelector ?? throw new ArgumentNullException(nameof(tfmSelector));
+            _upgradeOptions = upgradeOptions ?? throw new ArgumentNullException(nameof(upgradeOptions));
+
+            if (_upgradeOptions.Value is null)
+            {
+                throw new ArgumentException("Missing UpgradeReadinessOptions");
+            }
         }
 
         protected override async Task<bool> IsApplicableImplAsync(IUpgradeContext context, CancellationToken token)
@@ -200,14 +209,35 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Solution
 
         private async Task<bool> RunChecksAsync(IProject project, CancellationToken token)
         {
+            var upgradeGuidanceMessages = new List<string>();
             foreach (var check in _checks)
             {
                 Logger.LogTrace("Running readiness check {Id}", check.Id);
 
-                if (!await check.IsReadyAsync(project, token).ConfigureAwait(false))
+                var readiness = await check.IsReadyAsync(project, _upgradeOptions.Value, token).ConfigureAwait(false);
+                if (readiness == UpgradeReadiness.NotReady)
                 {
                     return false;
                 }
+                else if (!_upgradeOptions.Value.IgnoreUnsupportedFeatures && readiness == UpgradeReadiness.Unsupported)
+                {
+                    // an unsupported area has been detected. Capture a message to explain how to proceed.
+                    upgradeGuidanceMessages.Add(check.UpgradeMessage);
+                }
+            }
+
+            if (upgradeGuidanceMessages.Any())
+            {
+                // NOTE: the hidden assumption is that the UpgradeGuidance messages that were collected will be printed
+                // to console by the IUpgradeReadyCheck. We make these messages accessible by property so that they can be integrated into a UI in the future
+                // but assume that the IUpgradeReadyCheck displays them via logging because semantic logging provides color highlighting that we want to preserve.
+                Logger.LogError($"Project {{Name}} uses feature(s) that are not supported.{Environment.NewLine}If you would like to upgrade-assistant to continue anways please use the \"{{UnsupportedFeatures}}\" option.",
+                    project.FileInfo,
+                    "--ignore-unsupported-features");
+
+                // user has been informed about how to proceed.
+                // This project is not ready until the option '--ignore-unsupported-features' is provided.
+                return false;
             }
 
             return true;